阶段2收口与工程化：监控UI、验收脚本、pytest、CI、Docker、文档与bcrypt/Pydantic修复
- 监控：圆环百分比、流量曲线图、远程一键安装psutil(优先apt)
- 认证：passlib改为bcrypt；Pydantic改用ConfigDict
- 测试：scripts/smoke.py；backend pytest；User.id改为Integer
- CI：.github/workflows/ci.yml
- Docker：Dockerfile、docker-compose、docs/deploy.md
- 文档：design-monitor、plan、testing-steps 更新

Made-with: Cursor

+++ backend/models/user.py
@@ -1,6 +1,6 @@
 """用户模型。"""
 from datetime import datetime
-from sqlalchemy import String, DateTime, BigInteger
+from sqlalchemy import String, DateTime, Integer
 from sqlalchemy.orm import Mapped, mapped_column
 
 from database import Base
@@ -9,7 +9,7 @@ from database import Base
 class User(Base):
     __tablename__ = "users"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     username: Mapped[str] = mapped_column(String(64), unique=True, nullable=False, index=True)
     password_hash: Mapped[str] = mapped_column(String(255), nullable=False)
     created_at: Mapped[datetime] = mapped_column(DateTime, default=datetime.utcnow)
